# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BillingInterval": {
        "enum": [
          "month",
          "year",
          "one_time"
        ],
        "title": "BillingInterval",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanCreate": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_-]+$",
            "title": "Code",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/PlanFeatureIn"
            },
            "title": "Features",
            "type": "array"
          },
          "interval": {
            "$ref": "#/components/schemas/BillingInterval",
            "default": "month"
          },
          "is_public": {
            "default": true,
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_cents": {
            "minimum": 0.0,
            "title": "Price Cents",
            "type": "integer"
          },
          "provider_refs": {
            "additionalProperties": true,
            "title": "Provider Refs",
            "type": "object"
          },
          "trial_days": {
            "default": 0,
            "minimum": 0.0,
            "title": "Trial Days",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "name",
          "price_cents"
        ],
        "title": "PlanCreate",
        "type": "object"
      },
      "PlanFeatureIn": {
        "properties": {
          "credit_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Amount"
          },
          "feature_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Feature Key",
            "type": "string"
          },
          "is_hard_limit": {
            "default": true,
            "title": "Is Hard Limit",
            "type": "boolean"
          },
          "limit_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Value"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "feature_key"
        ],
        "title": "PlanFeatureIn",
        "type": "object"
      },
      "PlanFeatureOut": {
        "properties": {
          "credit_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Amount"
          },
          "feature_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Feature Key",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_hard_limit": {
            "default": true,
            "title": "Is Hard Limit",
            "type": "boolean"
          },
          "limit_value": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Value"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "feature_key",
          "id"
        ],
        "title": "PlanFeatureOut",
        "type": "object"
      },
      "PlanResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/PlanFeatureOut"
            },
            "title": "Features",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interval": {
            "$ref": "#/components/schemas/BillingInterval"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_cents": {
            "title": "Price Cents",
            "type": "integer"
          },
          "trial_days": {
            "title": "Trial Days",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "code",
          "name",
          "description",
          "price_cents",
          "currency",
          "interval",
          "trial_days",
          "is_public",
          "is_active",
          "features"
        ],
        "title": "PlanResponse",
        "type": "object"
      },
      "PlanUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "is_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Public"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "price_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Cents"
          },
          "provider_refs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Refs"
          }
        },
        "title": "PlanUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "product-platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "operationId": "list_plans_api_v1_plans_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanResponse"
                  },
                  "title": "Response List Plans Api V1 Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plans",
        "tags": [
          "plans"
        ]
      },
      "post": {
        "operationId": "create_plan_api_v1_plans_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/plans/{code}": {
      "patch": {
        "operationId": "update_plan_api_v1_plans__code__patch",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Plan",
        "tags": [
          "plans"
        ]
      }
    },
...
  }
}